requirements.txt
langchain
langchain-community
chromadb
gradio
python-dotenv
sentence-transformers
